Rename server actions util
